Huge update!
Started implementing the game mechanics. Basic behavior like
clicking/touching and changing pieces across the board.

There is a next colors bar to indicate the next colors and a score. The
score and the game over were not implemented.

Beisdes, this game is ugly as hell. It needs a menu, different playing
modes (?) and something like a bonus to redraw the next colors.